feat(sofle/eam): left hand encoder scrolling
Made it so that my left-hand encoder is a scroll wheel. The right-hand
encoder is dual purpose. It is a volume knob by default and media keys
when lowered.

This feels pretty natural. My hand is already on my left side when I'm
using my trackball. It also lets me use the trackball while scrolling